Added some basic time functionality.

+++ socket.hpp
@@ -1,6 +1,6 @@
 //Socket Header
 //	Created By:		Mike Moss
-//	Modified On:	04/25/2013
+//	Modified On:	05/19/2013
 
 //Required Libraries:
 //	wsock32 (windows only)
@@ -117,10 +117,10 @@ namespace msl
 			int check() const;
 
 			//Connection Timeout Mutator
-			void set_timeout(const unsigned int time_out);
+			void set_timeout(const long time_out);
 
 			//Connection Timeout Accessor
-			unsigned int timeout() const;
+			long timeout() const;
 
 			//IP Address Accessor (Read Only)
 			msl::ipv4 ip() const;
@@ -136,7 +136,7 @@ namespace msl
 			msl::ipv4 _address;
 			SOCKET _socket;
 			bool _hosting;
-			unsigned int _time_out;
+			long _time_out;
 	};
 
 	//Socket Class Stream Operator (Templated Function)
@@ -157,28 +157,28 @@ namespace msl
 }
 
 //Socket Create Function (Hosts a Socket Locally)
-SOCKET socket_create(const msl::ipv4 ip,const unsigned int time_out=0,const bool UDP=false,const unsigned int buffersize=200);
+SOCKET socket_create(const msl::ipv4 ip,const long time_out=0,const bool UDP=false,const unsigned int buffersize=200);
 
 //Socket Connection Function (Connects to a Remote Socket)
-SOCKET socket_connect(const msl::ipv4 ip,const unsigned int time_out=0,const bool UDP=false);
+SOCKET socket_connect(const msl::ipv4 ip,const long time_out=0,const bool UDP=false);
 
 //Socket Accept Function (Accepts a Remote Connection to a Local Socket)
-SOCKET socket_accept(const SOCKET socket,msl::ipv4& client_ip,const unsigned int time_out=0);
+SOCKET socket_accept(const SOCKET socket,msl::ipv4& client_ip,const long time_out=0);
 
 //Socket Close Function (Closes a Local Socket)
 SOCKET socket_close(const SOCKET socket);
 
 //Socket Check Read Function (Checks How Many Bytes there are to be Read, -1 on Error)
-int socket_check_read(const SOCKET socket,const unsigned int time_out=0);
+int socket_check_read(const SOCKET socket,const long time_out=0);
 
 //Socket Peek Function (Same as socket_read but Leaves Bytes in Socket Buffer)
-int socket_peek(const SOCKET socket,void* buffer,const unsigned int size,const unsigned int time_out=200,const int flags=0);
+int socket_peek(const SOCKET socket,void* buffer,const unsigned int size,const long time_out=200,const int flags=0);
 
 //Socket Read Function (Returns Number of Bytes Read, -1 on Error)
-int socket_read(const SOCKET socket,void* buffer,const unsigned int size,const unsigned int time_out=200,const int flags=0);
+int socket_read(const SOCKET socket,void* buffer,const unsigned int size,const long time_out=200,const int flags=0);
 
 //Socket Write Function (Returns Number of Bytes Sent, -1 on Error)
-int socket_write(const SOCKET socket,void* buffer,const unsigned int size,const unsigned int time_out=200,const int flags=0);
+int socket_write(const SOCKET socket,void* buffer,const unsigned int size,const long time_out=200,const int flags=0);
 
 //End Define Guards
 #endif
